move add-comp to shared-dir
